fix(ecs): pin archetype + world to force inline test enumeration
E5b added a new analysis frontier (SystemScheduler → World →
ensureComponentRegistered → Registry → FieldKind) that surfaced a
latent compile error in tests/ecs/archetype.zig's inline test (Tag
field used u8, which FieldKind.fromZigType rejects until M0.2
RTTI). The inline test was silently skipped before E5b because
core_tests' lazy-analysis frontier did not reach ecs.archetype or
ecs.world. Add the missing pins in src/core/root.zig (same pattern
as ecs.entity / ecs.tick / ecs.change_detection / ecs.scheduler
fixed in earlier milestones) and switch the test's Tag field to
u32 so the FieldKind whitelist accepts it. Behavioral semantics
of the inline test (sorted component_ids invariant) preserved.

+++ src/core/ecs/archetype.zig
@@ -450,7 +450,11 @@ test "Archetype init pins sorted component_ids and registry-driven sizes/aligns"
     defer reg.deinit(gpa);
 
     const Health = extern struct { current: f32 = 0, max: f32 = 100 };
-    const Tag = extern struct { v: u8 = 0 };
+    // M0.1 / E5b note: Tag uses `u32` rather than `u8` because the
+    // E4 `FieldKind` registry whitelist does not include `u8`
+    // (RTTI cleanup is M0.2). The test only cares that two
+    // components with distinct sizes/aligns sort correctly.
+    const Tag = extern struct { v: u32 = 0 };
 
     const id_h = try reg.registerComponent(gpa, Health);
     const id_t = try reg.registerComponent(gpa, Tag);
